requirements.txt
Flask
Gunicorn
Pillow
numpy
werkzeug
flask_wtf
matplotlib
pytest
flask_testing